requirements.txt
discord.py==2.5.2
yt_dlp==2025.3.31
pydub==0.25.1
PyNaCl==1.5.0